// angular/blog/src/app/types/todo.type.ts
export interface TODO_TYPE {
  userId: number;
  id: number;
  title: string;
  completed: boolean;
};

// interface TodoOptions {
//   hide: boolean;
// }

// export interface TODO extends TodoOptions {
//   userId: number;
//   id: number;
//   title: string;
//   completed: boolean;

//   print: () => boolean
// }

// class TODOItem implements TODO {
//   print() {
//     return false
//   }

//   userId = 0
//   id = 10
//   title = 'cd'
//   completed = false;
//   hide = true
// }

// class TODOSecondItem {}